requirements.txt
matplotlib.pyplot
matplotlib.patches
scipy
numpy
pandas
random
IPython.display
os